// admin/models/api.ts
export interface CreateStateRequest {
  name: string;
  uf: string;
}

export interface StateResponse {
  id: number;
  name: string;
  uf: string;
  createdAt: Date;
  updatedAt: Date;
}

export interface UpdateStateRequest {
  name: string;
  uf: string;
}
// ...
export interface PagedResponse<T> {
  page: number;
  size: number;
  totalElements: number;
  totalPages: number;
  data: T[];
}